+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
 ## [unreleased]
+
+## v0.3.0 - 2025-08-11
 ### Features
  - Summary command lists total duration and average calories per workout
  - List number of sport type in time range in summary